+++ tests/stubs/views/componentWithVariables.blade.php
@@ -1,0 +1,16 @@
+@php
+    $user = 'John';
+@endphp
+
+<card :title="$user">
+    My content
+</card>
+
+
+@php
+    $user = 'Jane';
+@endphp
+
+<card :title="$user">
+    My content
+</card>