+++ tests/stubs/views/globalFunction.blade.php
@@ -1,0 +1,7 @@
+@php
+$title = 'my title'
+@endphp
+
+<card :title="str_after(ucfirst($title), 'My ')">
+    {{ $title }}
+</card>