+++ tests/stubs/views/invalidComponent.blade.php
@@ -1,0 +1,3 @@
+<card ::::::invalid= class="{{ $type }}">
+    {{ $message }}
+</card>