+++ tests/stubs/views/selfClosingComponent.blade.php
@@ -1,0 +1,3 @@
+<?php $message = 'my message' ?>
+
+<alert type="error" :message="$message" />