+++ tests/stubs/views/templates/alert.blade.php
@@ -1,3 +1,0 @@
-<?php $message = 'my message' ?>
-
-<alert type="error" message="{{ $message }}" />