+++ tests/stubs/views/templates/dynamicProfile.blade.php
@@ -1,7 +1,0 @@
-@php
-    $user = 'John';
-@endphp
-
-<card title="{{ $user }}">
-    My content
-</card>