+++ tests/stubs/views/twoComponents.blade.php
@@ -1,0 +1,11 @@
+<card title="Profile 1">
+    <form>
+        <text-field name="name" label="Name"></text-field>
+    </form>
+</card>
+
+<card title="Profile 2">
+    <form>
+        <text-field name="name" label="Name"></text-field>
+    </form>
+</card>