+++ tests/stubs/views/viewModel.blade.php
@@ -1,0 +1,9 @@
+@php
+    $countries = [
+        'be' => 'Belgium',
+        'fr' => 'France',
+        'nl' => 'The Netherlands',
+    ];
+@endphp
+
+<select-field name="countries" :options="$countries" selected="fr" />