+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `laravel-blade-x` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.1.2 - 2018-10-02
+
+- make sure a component is registered only once
+- make sure kebab-cased props get passed to components camelCased
+
+## 1.1.1 - 2018-10-01
+
+- add view models
+
+## 1.0.0 - 2018-10-01
 
 - initial release